Fixed signed-off-by-test to diff between master and PR
The previous attempt would attempt to look at all commits between a
certain 'first' commit and HEAD, which would take longer and longer over times
as commits were added. Also, it would fail because travis only checks
out to a depth of 50, so sooner or later, our 'first' commit may not
exist.

This uses the 'git cherry' command to find all the commits that differ
between master and the PR, and check that *those* commits have
'Signed-off-by'. This also has the advantage that merge commits are
ignored by 'git cherry'

Signed-off-by: Bradon Kanyid <rattboi@gmail.com>